python/requirements.txt
pytest
requests
Random-Word-Generator
python-dotenv
pytest-html
pytest-xdist